pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mocharpa"
version = "0.1.0"
description = "A modern, extensible RPA bot framework with OOP and functional-friendly APIs"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "RPA Framework Team"}]
keywords = ["rpa", "automation", "uiautomation", "robotic-process-automation", "mocharpa"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typing_extensions>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
browser = [
    "playwright>=1.40",
]
http = [
    "requests>=2.28",
]
excel = [
    "openpyxl>=3.1",
]
word = [
    "python-docx>=0.8",
]
database = [
    "sqlalchemy>=2.0",
]
yaml = [
    "pyyaml>=6.0",
]
plugins = [
    "mocharpa[excel,word]",
    "playwright>=1.40",
    "requests>=2.28",
    "openpyxl>=3.1",
    "python-docx>=0.8",
    "sqlalchemy>=2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mocharpa", "mocharpa.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
